Create main.dart
The contacts page can sync with your phone's contacts, and you can share your location to the number, or list of numbers, that is in the string in the _sendSMS method. However, right now the phone number passed as an argument to recipients is my own for testing (you can input yours for testing as well). Therefore, it needs to be made such that when the user selects contacts via checkboxes, each of those numbers is added to the recipient list, and they will receive an SMS with their location on Google Maps.